fix: refactor to use context based cache helper
