+++ packages/sync/src/conflicts/BaseLink.ts
@@ -1,7 +1,7 @@
 import { ApolloLink, NextLink, Operation, Observable } from "apollo-link";
 import { InMemoryCache } from "apollo-cache-inmemory";
 import { isMutation } from "../utils/helpers";
-import { getObjectFromCache } from "../offline/storage/defaultStorage";
+import { getObjectFromCache } from "../utils/cacheHelper";
 import { ObjectState } from ".";
 import { logger } from "../links/LocalDirectiveFilterLink";
 import { ConflictResolutionData } from "./strategies/ConflictResolutionData";
